Synchronize with potato-branch again: debian/changelog: update to 1.6.10 scripts/dpkg-divert.pl: reinstate last writeability-patch. Modified to not abort on ENOENT: this indicates a directory does not (yet) exist. If this happens don't try to rename. scripts/update-alternatives.pl: + update usage-warning to reflect the new --config option + Document some global variables + Switch back to auto-mode when removing a manually selected alternative dselect/{dselect.h,main.cc,pkglist.cc}: remove CAN_RESIZE stuff dselect/methlist.cc: don't abort if getch fails due to interrupted syscall dselect/pkglist.cc: don't abort if getch fails due to interrupted syscall dselect/baselist.cc: put debug-statement in sigwinchhandler and actually resize the current terminal before redisplaying the screen

+++ scripts/dpkg-divert.pl
@@ -199,9 +199,13 @@ sub checkrename {
     # (hopefully) wont overwrite anything. If it succeeds, we
     # assume a writable filesystem.
     foreach $file ($rsrc,$rdest) {
-	open (TMP, ">> ${file}.dpkg-devert.tmp") || &quit("error checking \`$file': $!");
+	open (TMP, ">> ${file}.dpkg-devert.tmp") || $! == NOENT ||
+		&quit("error checking \`$file': $!");
 	close TMP;
-	unlink ("${file}.dpkg-devert.tmp");
+	if ($1 == ENOENT) {
+		$dorename = 0;
+	} else
+		unlink ("${file}.dpkg-devert.tmp");
     }
     if (@ssrc && @sdest &&
         !($ssrc[0] == $sdest[0] && $ssrc[1] == $sdest[1])) {
